Tessellation Experiment
This implements a simple subdivision tesselation setup w/ multiple compute shaders,
for fun.

This was implemented using https://github.com/Aeva/dream_machine.  The tess*.cs.glsl
files are all variants of the same file to change out the input and output UAV names,
since I don't currently have a mechanism for overriding that in dream machine.

Running this program should result in a black background with a white circle drawn in
the center.